implemented gson
Former-commit-id: 1f4b430add46c65bb750fa1c3f62a207f74d8d04

+++ parser/src/main/java/yasp/Player.java
@@ -10,7 +10,7 @@ public class Player {
     public List<Integer> lh = new ArrayList<Integer>();
     public List<Integer> gold = new ArrayList<Integer>();
     public List<Integer> xp = new ArrayList<Integer>();
-    public double stuns;
+    public float stuns;
     public Player(String personaname, long steamid){
         this.personaname = personaname;
         this.steamid = steamid;
